+++ pkg/resources/nifi/pvc_test.go
@@ -44,3 +44,53 @@ func TestPVC(t *testing.T) {
 	assert.Equal(t, "value", p.ObjectMeta.Annotations["annotation"])
 	assert.Equal(t, "value", p.ObjectMeta.Labels["label"])
 }
+
+func TestStorageConfigPVCExists(t *testing.T) {
+	r := resources.Reconciler{
+		NifiCluster: &v1.NifiCluster{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "cluster",
+				Namespace: "namespace",
+			},
+			Spec: v1.NifiClusterSpec{},
+		},
+	}
+	rec := Reconciler{
+		Reconciler: r,
+	}
+
+	pvcList := []corev1.PersistentVolumeClaim{
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "pvc1",
+				Annotations: map[string]string{
+					"storageName": "FOO",
+				},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "pvc2",
+				Annotations: map[string]string{
+					"storageName": "TARGET",
+				},
+			},
+		},
+		{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "pvc3",
+				Annotations: map[string]string{
+					"storageName": "BAR",
+				},
+			},
+		},
+	}
+
+	exists, pvc := rec.storageConfigPVCExists(pvcList, "TARGET")
+	assert.True(t, exists)
+	assert.Equal(t, pvc, &pvcList[1])
+
+	exists, pvc = rec.storageConfigPVCExists(pvcList, "NONEXISTENT")
+	assert.False(t, exists)
+	assert.Nil(t, pvc)
+}